requirements.txt
python-telegram-bot[job-queue]==21.6
feedparser
beautifulsoup4
requests
python-dotenv
